4_manage_chat_history/requirements.txt
langchain
langchain-community
langchain_groq
python-dotenv
transformers